Fixing a11y issues
