+++ components/atoms/Card.js
@@ -31,7 +31,9 @@ export default function Card(props) {
       <ul className="flex items-center justify-center pb-4">
         {props.links.map((d, idx) => (
           <li key={idx} className="w-auto text-blue-800 text-sm px-3">
-            <Link href={d.url}>{d.text}</Link>
+            <a aria-label={d.text + " " + props.service} href={d.url}>
+              {d.text}
+            </a>
           </li>
         ))}
       </ul>
