+++ components/organisms/Header.js
@@ -43,7 +43,7 @@ export function Header({ bannerTitle, bannerText, breadcrumbItems }) {
 
   return (
     <>
-      <nav className="skip-main">
+      <nav className="hidden" aria-label="skip-main">
         <a
           id="skipToMainContent"
           className="bg-custom-blue-dark text-white py-1 px-2 hover:bg-gray-dark"
